Created a fork for my presentation request
Hello Roger, I sent you a message about 2/8 presentation but wanted to make it a bit easier for you by creating a fork. As mentioned in my email and message, I would like to be added to 2/8. I thought I could later follow up with a presentation on "transportation for older adults". Thank you so much. Sorry to bother you (hopefully the fork makes it a bit easier?). Take care, Soyeon Lee